requirements.txt
# server-build
flask
flask-cors
flask_sqlalchemy
flask_login
# server-tools
Werkzeug
gunicorn
requests
python-dotenv
# parser-tools
bs4
markdown
PyYAML
imageio
imageio-ffmpeg
mpmath
# application-tools
matplotlib
numba
numpy
scipy
wikipedia
# bot-requireds
selenium
geckodriver-autoinstaller